refactor: create method getAuthQueryInfos in ay-uth-token-service
