--- src/app/core/models/profile/find-profile-details-request.model.ts
@@ -1,6 +1,0 @@
-export class FindProfileDetailsRequest {
-	constructor(
-		public authToken: string,
-		public authUserId: string,
-	) {}
-}
