+++ src/app/core/models/token/auth-query-infos.model.ts
@@ -1,0 +1,9 @@
+import { HttpHeaders } from '@angular/common/http';
+
+export class AuthQueryInfos {
+	constructor(
+		public authToken: string,
+		public authUserId: string,
+		public headers: HttpHeaders,
+	) {}
+}
