+++ src/app/core/services/product.service.ts
@@ -1,19 +1,19 @@
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import { Injectable, inject } from '@angular/core';
 import { AddOrRemoveProductToWishlistRequest } from '@models/product/add-or-remove-product-to-wishlist-request.model';
 import { FindProductByIdRequest } from '@models/product/find-product-by-id-request.model';
 import { ProductDetailsResponse } from '@models/product/product-details-response.model';
 import { ProductResponse } from '@models/product/product-response.model';
 import { SuccessResponse } from '@models/response/success-response.model';
-import { LocalStorageService } from '@services/auth/local-storage.service';
 import { Observable } from 'rxjs';
+import { AuthTokenService } from './auth/auth-token.service';
 
 @Injectable({
 	providedIn: 'root',
 })
 export class ProductService {
 	private _httpClient: HttpClient = inject(HttpClient);
-	private _localStorageService: LocalStorageService = inject(LocalStorageService);
+	private _authTokenService: AuthTokenService = inject(AuthTokenService);
 
 	private _BASE_URL = 'http://localhost:8080/api/v1/product';
 
@@ -30,8 +30,7 @@ export class ProductService {
 	public addOrRemoveProductToWishlist(
 		addOrRemoveProductToWishlistRequest: AddOrRemoveProductToWishlistRequest,
 	): Observable<SuccessResponse<ProductDetailsResponse>> {
-		const authToken = this._localStorageService.getAuthToken() as string;
-		const headers = new HttpHeaders().set('Authorization', `Bearer ${authToken}`);
+		const { headers } = this._authTokenService.getAuthQueryInfos();
 
 		return this._httpClient.post<SuccessResponse<ProductDetailsResponse>>(
 			`${this._BASE_URL}/${addOrRemoveProductToWishlistRequest.productId}/wishlist/${addOrRemoveProductToWishlistRequest.wishlistId}`,
